ConDec-359: Delete ConfigRest interface to remove the amount of code
